moved modalcontent to launchmodal, adjusted modal background color to match spinner background
