+++ client/src/components/DataLoading.module.scss
@@ -1,5 +1,5 @@
 .DataLoading {
-  background: rgba(0, 0, 0, 0.85);
+  background: rgba(0, 0, 0, 0.7);
   position: fixed;
   top: 0;
   left: 0;
