--- client/src/components/ModalContent.module.scss
@@ -1,81 +1,0 @@
-.TopContainer {
-  display: flex;
-
-  img {
-    max-width: 150px;
-  }
-
-  .TopDetails {
-    padding-left: 30px;
-  }
-}
-
-.InnerCard {
-  width: 60%;
-  left: 20%;
-  padding: 50px;
-}
-
-.Svg {
-  position: absolute;
-  top: 20px;
-  right: 20px;
-  font-size: 1rem;
-  cursor: pointer;
-}
-
-.TopContainer {
-  .DetailsTitle {
-    text-transform: uppercase;
-    font-size: 0.8rem;
-  }
-
-  p {
-    margin-bottom: 5px;
-  }
-}
-
-.TopContainer,
-.BottomContainer {
-  h3 {
-    text-transform: uppercase;
-    font-size: 1.35rem;
-    margin-bottom: 15px;
-    font-weight: 300;
-    letter-spacing: 1px;
-  }
-
-  .DetailsTitle {
-    font-weight: 700;
-    margin-bottom: 0;
-  }
-}
-.BottomContainer {
-  padding-top: 20px;
-  border-top: 1px solid;
-  margin-top: 20px;
-}
-
-.failed {
-  &::after {
-    content: "";
-    display: inline-block;
-    height: 1rem;
-    width: 40px;
-    background-color: crimson;
-    margin-left: 10px;
-    vertical-align: middle;
-  }
-}
-
-.success {
-  &::after {
-    content: "";
-    display: inline-block;
-    height: 1rem;
-    width: 40px;
-    margin-left: 10px;
-    vertical-align: middle;
-    background-color: green;
-  }
-}
